This update is not functional. Updating for task 3.
Complete:
RF Sniffing is complete
Added multithreading in a java-like syntax

Half-way:
The GUI is very basic. So it is in compliance with what was specified
for the project, but most likely won't be able to go any farther.

Not complete:
Model is NOT complete. Still needs to handle the RF module firing an
event when triggered.
Local TCP is NOT complete. Local TCP is unfinished/buggy.